Revert "Created pif classes Fixed the tests"
This reverts commit 24f9becc
